+++ packages/@expo/metro-config/src/serializer/__tests__/tree-shaking.test.ts
@@ -1014,6 +1014,7 @@ it(`removes unused exports`, async () => {
           ],
           "paths": {},
           "reactClientReferences": [],
+          "reactServerReferences": [],
           "requires": [],
         },
         "originFilename": "index.js",