+++ packages/expo-router/src/rsc/server.ts
@@ -13,6 +13,8 @@ import type { PathSpec } from './path';
 
 declare let globalThis: {
   __EXPO_RSC_CACHE__?: Map<string, any>;
+  __webpack_chunk_load__: (id: string) => Promise<any>;
+  __webpack_require__: (id: string) => any;
 };
 
 type Config = any;